refactor(media): extract Elo calculator, comparison queries, and score management from comparisons service
Extract 3 modules from service.ts (1558 → 1257 LOC) into comparisons/lib/:
- elo-calculator.ts — ELO_K, expectedScore(), drawTierOutcome() with unit tests
- comparison-queries.ts — findExistingComparison(), listComparisonsForMedia(),
  listAllComparisons(), normalizePairOrder()
- score-management.ts — getOrCreateScore(), updateEloScores(),
  recalcDimensionElo(), recalcAllDimensions()

All 247 tests pass (236 existing + 11 new). [tb-369]
